refactor: Improve Logger classes file location and simplify imports
