+++ p2pfl/management/logger/logger.py
@@ -23,79 +23,11 @@ P2PFL Logger.
 
 """
 
-import datetime
 import logging
 from p2pfl.experiment import Experiment
 from typing import Any, Dict, Optional
 
 from p2pfl.management.metric_storage import GlobalLogsType, LocalLogsType
-
-#########################################
-#    Logging handler (transmit logs)    #
-#########################################
-
-
-class DictFormatter(logging.Formatter):
-    """Formatter (logging) that returns a dictionary with the log record attributes."""
-
-    def format(self, record):
-        """
-        Format the log record as a dictionary.
-
-        Args:
-            record: The log record.
-
-        """
-        # Get node
-        if not hasattr(record, "node"):
-            raise ValueError("The log record must have a 'node' attribute.")
-        log_dict = {
-            "timestamp": datetime.datetime.fromtimestamp(record.created),
-            "level": record.levelname,
-            "node": record.node,  # type: ignore
-            "message": record.getMessage(),
-        }
-        return log_dict
-
-
-
-
-#########################
-#    Colored logging    #
-#########################
-
-# COLORS
-GRAY = "\033[90m"
-RED = "\033[91m"
-YELLOW = "\033[93m"
-GREEN = "\033[92m"
-BLUE = "\033[94m"
-CYAN = "\033[96m"
-RESET = "\033[0m"
-
-
-class ColoredFormatter(logging.Formatter):
-    """Formatter that adds color to the log messages."""
-
-    def format(self, record):
-        """
-        Format the log record with color.
-
-        Args:
-            record: The log record.
-
-        """
-        # Warn level color
-        if record.levelname == "DEBUG":
-            record.levelname = BLUE + record.levelname + RESET
-        elif record.levelname == "INFO":
-            record.levelname = GREEN + record.levelname + RESET
-        elif record.levelname == "WARNING":
-            record.levelname = YELLOW + record.levelname + RESET
-        elif record.levelname == "ERROR" or record.levelname == "CRITICAL":
-            record.levelname = RED + record.levelname + RESET
-        return super().format(record)
-
 
 ###################
 #    Interface    #
